refactor: 테스트하기 어려운 도메인 로직 분리
