+++ src/domain/RacingCarGame.js
@@ -1,6 +1,3 @@
-const GAME_RULE = require('../constant/gameRule');
-const Random = require('../util/Random');
-
 class RacingCarGame {
   #cars;
 
@@ -8,22 +5,10 @@ class RacingCarGame {
     this.#cars = cars;
   }
 
-  moveCars() {
-    const carsMovingSuccesses = this.getCarsMovingSuccesses();
+  moveCars(carsMovingSuccesses) {
     this.#cars.forEach((car, index) => {
       if (carsMovingSuccesses[index]) car.move();
     });
-  }
-
-  getCarsMovingSuccesses() {
-    return Array.from(
-      { length: this.#cars.length },
-      () =>
-        Random.pickNumberInRange(
-          GAME_RULE.RANDOM_LOWER_INCLUSIVE,
-          GAME_RULE.RANDOM_UPPER_INCLUSIVE,
-        ) >= GAME_RULE.MOVING_CONDITION_NUMBER,
-    );
   }
 
   getCarsInfo() {
@@ -34,6 +19,10 @@ class RacingCarGame {
     const carsInfo = this.getCarsInfo();
     const maxDistance = Math.max(...carsInfo.map((carInfo) => carInfo.distance));
     return carsInfo.filter(({ distance }) => distance === maxDistance).map(({ name }) => name);
+  }
+
+  getCountOfCars() {
+    return this.#cars.length;
   }
 }
